wrapping tripadvisor's ratings image in thumbr
